Don't use Hash#symbolize_keys, which Sidekiq used to monkeypatch
We're all familiar with `Hash#symbolize_keys`, but it isn't part of Ruby.

We were relying on this in our code in `Workers::DependencyUpdater`, but this was being monkeypatched in by Sidekiq (although usually it'd be Rails doing this). [Sidekiq 5.0.1](https://github.com/mperham/sidekiq/blob/master/Changes.md#501) removed all of these monkeypatches, breaking our code.

This implements our own `symbolize_hash_keys` function for doing this.